# docs/swagger.yaml
definitions:
  models.Area:
    properties:
      id:
        type: integer
      manager:
        type: string
      name:
        type: string
    type: object
  models.Customer:
    properties:
      cccd:
        type: string
      id:
        type: integer
      keypadPass:
        type: string
      name:
        type: string
      phone:
        type: string
      rfidPass:
        type: string
      schedulers:
        items:
          $ref: "#/definitions/models.Scheduler"
        type: array
    required:
      - cccd
    type: object
  models.DeleteCustomer:
    properties:
      cccd:
        type: string
    required:
      - cccd
    type: object
...
  models.Doorlock:
    properties:
      description:
        type: string
      doorSerialId:
        type: string
      gatewayId:
        type: string
      id:
        type: integer
      lastOpenTime:
        type: integer
      location:
        type: string
      schedulers:
        items:
          $ref: "#/definitions/models.Scheduler"
        type: array
    type: object
...
  models.Gateway:
    properties:
      areaId:
        type: integer
      doorlocks:
        items:
          $ref: "#/definitions/models.Doorlock"
        type: array
      gatewayId:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Scheduler:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      customerId:
        type: string
      doorSerialId:
        type: string
      employeeId:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      id:
        type: integer
      lecturerId:
        type: string
      lecturerName:
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      studentId:
        type: string
      weekDay:
        type: integer
    type: object
...
  models.SwagCreateArea:
    properties:
      gateway:
        $ref: "#/definitions/models.Gateway"
      manager:
        type: string
      name:
        type: string
    type: object
  models.SwagCreateCustomer:
    properties:
      cccd:
        type: string
      keypadPass:
        type: string
      name:
        type: string
      phone:
        type: string
      rfidPass:
        type: string
    required:
      - cccd
    type: object
...
  models.SwagUpdateArea:
    properties:
      gateway:
        $ref: "#/definitions/models.Gateway"
      id:
        type: integer
      manager:
        type: string
      name:
        type: string
    type: object
...
  models.UserSchedulerReq:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      customerId:
        type: string
      doorSerialId:
        type: string
      doorlockId:
        type: string
      employeeId:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      gatewayId:
        type: string
      id:
        type: integer
      lecturerId:
        type: string
      lecturerName:
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      studentId:
        type: string
      weekDay:
        type: integer
    required:
      - doorlockId
      - gatewayId
    type: object
  utils.ErrorResponse:
    properties:
      errorMsg:
        type: string
      msg:
        type: string
      statusCode:
        type: integer
    type: object
info:
  contact: {}
paths:
  /v1/area:
    delete:
      consumes:
        - application/json
      description: Delete area using "id" field
      parameters:
        - description: Area ID
          in: body
          name: data
          required: true
          schema:
            allOf:
              - type: object
              - properties:
                  id:
                    type: integer
                type: object
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Delete Area By ID
    patch:
      consumes:
        - application/json
      description: Update area, must have "id" field
      parameters:
        - description: Fields need to update a area
          in: body
          name: data
          required: true
          schema:
            $ref: "#/definitions/models.SwagUpdateArea"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Update Area By ID
    post:
      consumes:
        - application/json
      description: Create area
      parameters:
        - description: Fields need to create a area
          in: body
          name: data
          required: true
          schema:
            $ref: "#/definitions/models.SwagCreateArea"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Area"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Create Area
  /v1/area/{id}:
    get:
      description: find area info by area id
      parameters:
        - description: Area ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Area"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Find Area By ID
  /v1/areas:
    get:
      description: find all areas info
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: "#/definitions/models.Area"
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Find All Area
  /v1/customer:
    delete:
      consumes:
        - application/json
      description: Delete customer using "cccd" field
      parameters:
        - description: Customer CCCD
          in: body
          name: data
          required: true
          schema:
            $ref: "#/definitions/models.DeleteCustomer"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Delete Customer By CCCD
    patch:
      consumes:
        - application/json
      description: Update customer, must have correct "id" and "cccd" field
      parameters:
        - description: Fields need to update a customer
          in: body
          name: data
          required: true
          schema:
            $ref: "#/definitions/models.Customer"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Update Customer By ID
    post:
      consumes:
        - application/json
      description: Create customer
      parameters:
        - description: Fields need to create a customer
          in: body
          name: data
          required: true
          schema:
            $ref: "#/definitions/models.SwagCreateCustomer"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Customer"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Create Customer
  /v1/customer/{cccd}:
    get:
      description: find customer info by customer cccd
      parameters:
        - description: Customer CCCD
          in: path
          name: cccd
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Customer"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Find Customer By CCCD
  /v1/customer/{cccd}/scheduler:
    post:
      consumes:
        - application/json
      description: Add scheduler that allows customer open specific door
      parameters:
        - description: Request with Scheduler, GatewayID, DoorlockID
          in: body
          name: data
          required: true
          schema:
            $ref: "#/definitions/models.UserSchedulerReq"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Add Door Open Scheduler For Customer
  /v1/customers:
    get:
      description: find all customers info
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: "#/definitions/models.Customer"
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/utils.ErrorResponse"
      summary: Find All Customer
...
swagger: "2.0"